Complete Deposit/Withdraw: - TODO needs some read model and move snapshotting out of aggregate process

+++ priv/repo/migrations/20161008182303_create_wallet.exs
@@ -4,7 +4,7 @@ defmodule SimplePay.Repo.Migrations.CreateWallet do
   def change do
     create table(:wallets) do
       add :user_id, references(:users, on_delete: :nothing)
-      add :balance, :float, default: 0.0
+      add :balance, :integer, default: 0
       add :last_event_processed, :integer, default: -1
 
       timestamps()
